// frontend/src/generated-api/model/userData.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * BS.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface UserData {
  id: string;
  email: string;
  username: string;
  isEmailConfirm: boolean;
  isProfileFilled: boolean;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  contactUrl?: string | null;
  /** @nullable */
  photoUrl?: string | null;
}